requirements/airflow.txt
apache-airflow==2.9.3
psycopg2-binary==2.9.9
requests==2.31.0
openai==1.55.0
httpx>=0.25.0,<0.28.0
tiktoken==0.8.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
pgvector==0.3.5
